Add recommended AMI upgrade
This is a mega-commit that's a rebase of weeks of messy temporary
commits.

This adds "upgrade recommendations" to AMIable. When you're looking at
an AMI you'll be told if there's an upgrade available and exactly
which AMI you should upgrade to.

Also pulls out lots of redundant functionality. This service is a
little more focused on AMIs now.