Adds play-googleauth to Amiable
Adds the required Google configuration properties and code. Includes
the play googleauth library as a dependency (dev version see below) to
provide Google auth functionality. It also adds the controller and
required routing / templates to support login and logout behaviours.

For dev mode, this includes an nginx setup script. The DEV config file
now needs to include Google credentials to run.

**Note**:
This uses the WIP version of play-googleauth. It serves to test that
pull request
(https://github.com/guardian/play-googleauth/pull/39). This should be
merged only after that is released, and with a proper published
version of that library as a dependency.