Automatically determine obsolete Ubuntu distros
Replaces hard-coded list of supported Ubuntu versions with logic to
work this out, based on the current date. This is done using
canonical's appraoch, which is that LTS releases are April in even
numbered years, with 5 year support. All other releases have just 9
months support.

Getting the current date from the template (to pass to the helper
function) is not a great pattern, but this is consistent with the
current approach in these templates.